Moved jsHexViewer to new tools project
