+++ pom.xml
@@ -34,6 +34,7 @@
   <modules>
     <module>core</module>
     <module>formats</module>
+    <module>tools</module>
   </modules>
 
   <scm>
